Fix issue with TryGetMember/ItemSerializationInfo.
These should return false if the contract is of the wrong type instead of
throwing an Exception.

This is because JsonSerializerAdapter is both an IBsonArraySerializer and
an IBsonDocumentSerializer and so either method may be called to test if
there's a member of item serializtion info, regardless of the contract
type.

This happens in StringFieldDefinitionHelper.Resolve when resolving a
field using dot notation.